+++ packages/gasket-log-plugin/test.js
@@ -39,6 +39,9 @@ describe('log plugin', function () {
       const gasket = {
         exec: async function exec() {
         },
+        command: {
+          id: 'bogus'
+        },
         config: {
           env: 'test',
           winston: {}
@@ -54,7 +57,9 @@ describe('log plugin', function () {
       const exec = sinon.stub().resolves();
       const gasket = {
         exec,
-        command: 'start',
+        command: {
+          id: 'start'
+        },
         config: {
           env: 'test'
         }
@@ -72,7 +77,9 @@ describe('log plugin', function () {
       const exec = sinon.stub().resolves(['bar', 'bazz', undefined, ['apple', 'banana'], [undefined]]);
       const gasket = {
         exec,
-        command: 'start',
+        command: {
+          id: 'start'
+        },
         config: {
           env: 'test',
           winston: {
@@ -94,7 +101,9 @@ describe('log plugin', function () {
         exec: async function exec() {
           return null;
         },
-        command: 'start',
+        command: {
+          id: 'start'
+        },
         config: {
           env: 'test'
         }
@@ -108,9 +117,64 @@ describe('log plugin', function () {
       await plugin.hooks.init.handler(gasket);
     });
 
+    it('sets local option for local env', async () => {
+      const gasket = {
+        exec: sinon.stub(),
+        command: {
+          id: 'start'
+        },
+        config: {
+          env: 'local'
+        }
+      };
+
+      const plugin = assumeLogInit(actual => {
+        assume(actual.local).true();
+      });
+
+      await plugin.hooks.init.handler(gasket);
+    });
+
+    it('sets local option for non-start commands', async () => {
+      const gasket = {
+        exec: sinon.stub(),
+        command: {
+          id: 'fake'
+        },
+        config: {
+          env: 'test'
+        }
+      };
+
+      const plugin = assumeLogInit(actual => {
+        assume(actual.local).true();
+      });
+
+      await plugin.hooks.init.handler(gasket);
+    });
+
+    it('supports older gasket.command format', async () => {
+      const gasket = {
+        exec: sinon.stub(),
+        command: 'start',
+        config: {
+          env: 'test'
+        }
+      };
+
+      const plugin = assumeLogInit(actual => {
+        assume(actual.local).false();
+      });
+
+      await plugin.hooks.init.handler(gasket);
+    });
+
     it('forces { exitOnError: true }', async () => {
       const gasket = {
         exec: async function exec() {
+        },
+        command: {
+          id: 'bogus'
         },
         config: {
           winston: { exitOnError: false }