+++ packages/gasket-metadata-plugin/lib/index.js
@@ -44,6 +44,16 @@ module.exports = {
         // eslint-disable-next-line require-atomic-updates
         plugins[idx] = pluginData;
       });
+
+      // Loading preset metadata from module
+      presets.forEach(preset => {
+        if (preset.module.metadata) {
+          Object.keys(preset.module.metadata).reduce((acc, cur) => {
+            acc[cur] = preset.module.metadata[cur];
+            return acc;
+          }, preset);
+        }
+      });
     },
     metadata(gasket, pluginData) {
       return {