+++ packages/gasket-start-plugin/lib/index.js
@@ -1,0 +1,10 @@
+const create = require('./create');
+const getCommands = require('./get-commands');
+
+module.exports = {
+  name: 'start',
+  hooks: {
+    create,
+    getCommands
+  }
+};